nautilus_live::config

Struct LiveNodeConfig

pub struct LiveNodeConfig {

    pub environment: Environment,
    pub trader_id: TraderId,
    pub load_state: bool,
    pub save_state: bool,
    pub logging: LoggerConfig,
    pub instance_id: Option<UUID4>,
    pub timeout_connection: u32,
    pub timeout_reconciliation: u32,
    pub timeout_portfolio: u32,
    pub timeout_disconnection: u32,
    pub timeout_post_stop: u32,
    pub timeout_shutdown: u32,
    pub cache: Option<CacheConfig>,
    pub msgbus: Option<MessageBusConfig>,
    pub portfolio: Option<PortfolioConfig>,
    pub streaming: Option<StreamingConfig>,
    pub data_engine: LiveDataEngineConfig,
    pub risk_engine: LiveRiskEngineConfig,
    pub exec_engine: LiveExecEngineConfig,
    pub data_clients: HashMap<String, LiveDataClientConfig>,
    pub exec_clients: HashMap<String, LiveExecClientConfig>,

}

Configuration for live Posei system nodes.

Fields
environment: Environment

The trading environment.

trader_id: TraderId

The trader ID for the node.

load_state: bool

If trading strategy state should be loaded from the database on start.

save_state: bool

If trading strategy state should be saved to the database on stop.

logging: LoggerConfig

The logging configuration for the kernel.

instance_id: Option<UUID4>

The unique instance identifier for the kernel

timeout_connection: u32

The timeout (seconds) for all clients to connect and initialize.

timeout_reconciliation: u32

The timeout (seconds) for execution state to reconcile.

timeout_portfolio: u32

The timeout (seconds) for portfolio to initialize margins and unrealized pnls.

timeout_disconnection: u32

The timeout (seconds) for all engine clients to disconnect.

timeout_post_stop: u32

The timeout (seconds) after stopping the node to await residual events before final shutdown.

timeout_shutdown: u32

The timeout (seconds) to await pending tasks cancellation during shutdown.

cache: Option<CacheConfig>

The cache configuration.

msgbus: Option<MessageBusConfig>

The message bus configuration.

portfolio: Option<PortfolioConfig>

The portfolio configuration.

streaming: Option<StreamingConfig>

The configuration for streaming to feather files.

data_engine: LiveDataEngineConfig

The live data engine configuration.

risk_engine: LiveRiskEngineConfig

The live risk engine configuration.

exec_engine: LiveExecEngineConfig

The live execution engine configuration.

data_clients: HashMap<String, LiveDataClientConfig>

The data client configurations.

exec_clients: HashMap<String, LiveExecClientConfig>

The execution client configurations.